// frontend/package/src/interfaces/nextOfKin.ts
export interface NextOfKin {
  id: number;

  member: number;

  member_number: string;

  first_name: string;

  other_names: string;

  relationship: string;

  national_id: string;

  phone_number: string;

  physical_address: string;

  is_primary: boolean;

  created_at: string;

  updated_at: string;
}